server/requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.0.3
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
requests==2.32.3
urllib3==2.2.3
Werkzeug==3.0.4
youtube-transcript-api==0.6.2
zipp==3.20.2
livekit >= 0.15.0
livekit-protocol
livekit-agents>=0.10.0
livekit-plugins-openai>=0.10.0
aiodns==3.2.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2
async-timeout==4.0.3
attrs==24.2.0
av==13.1.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.0.3
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
livekit==0.17.5
livekit-agents==0.10.2
livekit-api==0.7.1
livekit-plugins-openai==0.10.4
livekit-protocol==0.6.0
MarkupSafe==3.0.1
multidict==6.1.0
openai==1.51.2
pillow==10.3.0
propcache==0.2.0
protobuf==5.28.2
psutil==5.9.8
pycares==4.4.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
PyJWT==2.9.0
python-dotenv==1.0.1
pytube==15.0.0
requests==2.32.3
sniffio==1.3.1
tqdm==4.66.5
types-protobuf==4.25.0.20240417
typing_extensions==4.12.2
urllib3==2.2.3
watchfiles==0.24.0
Werkzeug==3.0.4
yarl==1.15.2
youtube-transcript-api==0.6.2
zipp==3.20.2
